+++ src/index.ts
@@ -1,10 +1,4 @@
 export { createRouter, Router } from "./router";
 export { querySerializers } from "./query-serializers";
-export type {
-  RouteDefinition,
-  QuerySerializer,
-  QuerySerializers,
-  SchemaValidator,
-  InferParams,
-  InferQuery
-} from "./types";
+export type { RouteDefinition, QuerySerializer } from "./router";
+export type { QuerySerializers } from "./query-serializers";